Move #[godot_dyn] code generation to its own function
